Move LoadEditorWorld to WorldCreator
