+++ Assets/GothicVR/Scripts/Importer/PhoenixImporter.cs
@@ -8,15 +8,9 @@ using GVR.Util;
 using PxCs.Interface;
 using System;
 using System.IO;
-using System.Linq;
 using TMPro;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using UnityEngine.TextCore.LowLevel;
-
-#if UNITY_EDITOR
-using UnityEditor.SceneManagement;
-#endif
 
 namespace GVR.Importer
 {
@@ -88,53 +82,6 @@ namespace GVR.Importer
         {
             SingletonBehaviour<WorldCreator>.GetOrCreate().LoadWorld(vdfPtr, "world", "ENTRANCE_SURFACE_OLDMINE");
         }
-
-#if UNITY_EDITOR
-        /// <summary>
-        /// Loads the world.
-        /// </summary>
-        /// <param name="vdfPtr">The VDF pointer.</param>
-        /// <param name="zen">The name of the .zen world to load.</param>
-        public void LoadEditorWorld(IntPtr vdfPtr, string zen)
-        {
-            var worldScene = EditorSceneManager.GetSceneByName(zen);
-
-            if (!worldScene.isLoaded)
-            {
-                // unload the current scene and load the new one
-                EditorSceneManager.UnloadSceneAsync(EditorSceneManager.GetActiveScene());
-                EditorSceneManager.LoadScene(zen, LoadSceneMode.Additive);
-                worldScene = EditorSceneManager.GetSceneByName(zen); // we do this to reload the values for the new scene which are no updated for the above cast
-            }
-
-            var world = WorldBridge.LoadWorld(vdfPtr, $"{zen}.zen"); // world.zen -> G1, newworld.zen/oldworld.zen/addonworld.zen -> G2
-
-            PhoenixBridge.VdfsPtr = vdfPtr;
-            PhoenixBridge.World = world;
-
-            var worldGo = new GameObject("World");
-
-            // We use SampleScene because it contains all the VM pointers and asset cache necesarry to generate the world
-            var sampleScene = EditorSceneManager.GetSceneByName("SampleScene");
-            EditorSceneManager.SetActiveScene(sampleScene);
-            sampleScene.GetRootGameObjects().Append(worldGo);
-
-            var worldMesh = SingletonBehaviour<MeshCreator>.GetOrCreate().Create(world, worldGo);
-            // SingletonBehaviour<VobCreator>.GetOrCreate().Create(worldGo, world);
-            // SingletonBehaviour<WaynetCreator>.GetOrCreate().Create(worldGo, world);
-            // SingletonBehaviour<WorldCreator>.GetOrCreate().PostCreate(worldMesh);
-
-            // SingletonBehaviour<DebugAnimationCreator>.GetOrCreate().Create();
-
-            // move the world to the correct scene
-            EditorSceneManager.MoveGameObjectToScene(worldGo, worldScene);
-
-            // Subscribe the SetActiveScene method to the sceneLoaded event
-            // so that we can set the proper scene as active when the scene is finally loaded
-            // is related to occlusion culling
-            EditorSceneManager.sceneLoaded += (scene, mode) => EditorSceneManager.SetActiveScene(scene);
-        }
-#endif
 
         private void LoadGothicVM(string G1Dir)
         {
